feat!: IEnvironmentReporter now reports nullable values (#143)
Previously, we reported "unknown" from `IEnvironmentReporter`
implementation when properties couldn't be fetched.

This is problematic since "unknown" could be part of the valid domain of
values. Instead, let's follow the LD default of omitting properties
entirely.

To achieve this, the structs returned by `IEnvironmentReporter` are now
nullable.

Downstream in the SDK context decorators, we can check for null and
omit.